agri_chatbot/requirements.txt
image.pngimage.pngimage.pngflask>=2.3.0
python-dotenv>=1.0.0
openai>=1.3.0
requests>=2.31.0
google-generativeai>=0.3.0
rapidfuzz>=3.0.0
deep-translator>=1.11.4
pyttsx3>=2.90 